Reorganizing code
